rustc_hir: refactor rustc_hir_id re-exports
Removes the `pub use rustc_hir_id as hir_id` re-export in `rustc_hir`. This was rarely used and bloated import suggestions a lot.
