+++ src/tools/clippy/clippy_lints/src/cargo/mod.rs
@@ -9,7 +9,7 @@ use clippy_config::Conf;
 use clippy_utils::diagnostics::span_lint;
 use clippy_utils::is_lint_allowed;
 use rustc_data_structures::fx::FxHashSet;
-use rustc_hir::hir_id::CRATE_HIR_ID;
+use rustc_hir::CRATE_HIR_ID;
 use rustc_lint::{LateContext, LateLintPass, Lint};
 use rustc_session::impl_lint_pass;
 use rustc_span::DUMMY_SP;
