+++ src/tools/clippy/clippy_lints/src/only_used_in_recursion.rs
@@ -5,9 +5,9 @@ use core::cell::Cell;
 use rustc_data_structures::fx::FxHashMap;
 use rustc_errors::Applicability;
 use rustc_hir::def_id::DefId;
-use rustc_hir::hir_id::HirIdMap;
 use rustc_hir::{
-    Body, Expr, ExprKind, HirId, ImplItem, ImplItemImplKind, ImplItemKind, Node, PatKind, TraitItem, TraitItemKind,
+    Body, Expr, ExprKind, HirId, HirIdMap, ImplItem, ImplItemImplKind, ImplItemKind, Node, PatKind,
+    TraitItem, TraitItemKind,
 };
 use rustc_lint::{LateContext, LateLintPass};
 use rustc_middle::ty::{self, ConstKind, GenericArgKind, GenericArgsRef};
